Comments on the back of recent CRAN submission - added return fields
Thanks,

Please provide a link to the used webservices to the description field of your DESCRIPTION file in the form <http:...> or <[https:...]https:...> with angle brackets for auto-linking and no space after 'http:' and 'https:'.

Please add \value to .Rd files regarding exported methods and explain the functions results in the documentation. Please write about the structure of the output (class) and also what the output means. (If a function does not return a value, please document that too, e.g.
\value{No return value, called for side effects} or similar) Missing Rd-tags:
      left_xl.Rd: \value
      len_xl.Rd: \value
      linkScrapeR.Rd: \value
      mid_xl.Rd: \value
      nhs_data_elements.Rd: \value
      right_xl.Rd: \value
      scrapeR.Rd: \value
      tableR.Rd: \value

Please fix and resubmit.

Best,
Gregor Seyer